WPCS - Check for Groups - Membership
On saving, WPCS is run, so there are some spacing changes. More importantly, this installation was new and there aren't yet any groups and as an Admin, I didn't assign myself a level.

In the common plugin run a check for group

 `		$bp_groups = groups_get_total_group_count();
		if ( null === $bp_groups ) {`

and check for restriction

`	if ( null !== $pmpro_bp_options['pmpro_bp_restrictions'] ) {`